@@ -4,17 +4,18 @@
   "private": true,
   "homepage": "https://github.com/Irchik512/goit-react-hw-06-phonebook",
   "dependencies": {
-    "@reduxjs/toolkit": "^1.8.1",
+    "@reduxjs/toolkit": "^1.9.3",
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
-    "nanoid": "^3.2.0",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "react-redux": "^8.0.1",
-    "react-scripts": "5.0.0",
+    "nanoid": "^3.3.6",
+    "react": "^18.1.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-scripts": "5.0.1",
     "redux-devtools-extension": "^2.13.9",
-    "styled-components": "^5.3.3",
+    "redux-persist": "^6.0.0",
+    "styled-components": "^5.3.9",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
