@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import { Contact } from 'components/ContactList/ContactsListItem.styled';
 import { Button } from 'components/ContactForm/ContactForm.styled';
 import { useDispatch } from 'react-redux';
-import { remove } from '../../redux';
+import { removeContact } from 'redux/contactsSlice';
 
 export default function ContactsListItem({ name, number, id }) {
   const dispatch = useDispatch();
@@ -12,7 +12,7 @@ export default function ContactsListItem({ name, number, id }) {
       <p>
         {name}: {number}
       </p>
-      <Button type="button" onClick={() => dispatch(remove(id))}>
+      <Button type="button" onClick={() => dispatch(removeContact(id))}>
         Delete
       </Button>
     </Contact>
