@@ -1,12 +1,34 @@
-import { configureStore } from '@reduxjs/toolkit';
-import { contactsSlice } from './itemsSlice';
-import { filterSlice } from './filterSlice';
+import { combineReducers, configureStore } from '@reduxjs/toolkit';
+import { persistStore, persistReducer } from 'redux-persist';
+import storage from 'redux-persist/lib/storage';
+import { contactsReduser } from 'redux/contactsSlice';
+import { filterReduser } from 'redux/filterSlice';
+
+const rootReducer = combineReducers({
+  contacts: contactsReduser,
+  filter: filterReduser,
+});
+
+const persistConfig = {
+  key: 'root',
+  storage: storage,
+  whitelist: ['contacts'],
+};
+const persistedReducer = persistReducer(persistConfig, rootReducer);
 
 export const store = configureStore({
-  reducer: {
-    contacts: {
-      items: contactsSlice.reducer,
-      filter: filterSlice.reducer,
-    },
-  },
+  reducer: persistedReducer,
+  // middleware: getDefaultMiddleware =>
+  //   getDefaultMiddleware({
+  //     serializableCheck: {
+  //       // Ignore these action types
+  //       ignoredActions: ['your/action/type'],
+  //       // Ignore these field paths in all actions
+  //       ignoredActionPaths: ['meta.arg', 'payload.timestamp'],
+  //       // Ignore these paths in the state
+  //       ignoredPaths: ['items.dates'],
+  //     },
+  //   }),
 });
+
+export const persistor = persistStore(store);
